Person: added JobAssignment to fk check(s)

+++ src/main/java/com/mcommings/campaigner/services/people/PersonService.java
@@ -5,6 +5,7 @@ import com.mcommings.campaigner.models.RepositoryHelper;
 import com.mcommings.campaigner.models.people.Person;
 import com.mcommings.campaigner.repositories.IWealthRepository;
 import com.mcommings.campaigner.repositories.people.IAbilityScoreRepository;
+import com.mcommings.campaigner.repositories.people.IJobAssignmentRepository;
 import com.mcommings.campaigner.repositories.people.IPersonRepository;
 import com.mcommings.campaigner.repositories.people.IRaceRepository;
 import jakarta.transaction.Transactional;
@@ -18,6 +19,7 @@ import java.util.Arrays;
 import java.util.List;
 
 import static com.mcommings.campaigner.enums.ErrorMessage.*;
+import static com.mcommings.campaigner.enums.ForeignKey.FK_PERSON;
 import static java.util.Objects.isNull;
 
 @Service
@@ -27,14 +29,17 @@ public class PersonService implements IPerson {
     private final IRaceRepository raceRepository;
     private final IWealthRepository wealthRepository;
     private final IAbilityScoreRepository abilityScoreRepository;
+    private final IJobAssignmentRepository jobAssignmentRepository;
 
     @Autowired
     public PersonService(IPersonRepository personRepository, IRaceRepository raceRepository,
-                         IWealthRepository wealthRepository, IAbilityScoreRepository abilityScoreRepository) {
+                         IWealthRepository wealthRepository, IAbilityScoreRepository abilityScoreRepository,
+                         IJobAssignmentRepository jobAssignmentRepository) {
         this.personRepository = personRepository;
         this.raceRepository = raceRepository;
         this.wealthRepository = wealthRepository;
         this.abilityScoreRepository = abilityScoreRepository;
+        this.jobAssignmentRepository = jobAssignmentRepository;
     }
 
     @Override
@@ -65,7 +70,9 @@ public class PersonService implements IPerson {
         if (RepositoryHelper.cannotFindId(personRepository, personId)) {
             throw new IllegalArgumentException(DELETE_NOT_FOUND.message);
         }
-        //TODO: fk check when Person is a fk
+        if (RepositoryHelper.isForeignKey(getReposWherePersonIsAForeignKey(), FK_PERSON.columnName, personId)) {
+            throw new DataIntegrityViolationException(DELETE_FOREIGN_KEY.message);
+        }
         personRepository.deleteById(personId);
     }
 
@@ -100,6 +107,10 @@ public class PersonService implements IPerson {
 
     private boolean personAlreadyExists(Person person) {
         return personRepository.personExists(person).isPresent();
+    }
+
+    private List<CrudRepository> getReposWherePersonIsAForeignKey() {
+        return new ArrayList<>(Arrays.asList(jobAssignmentRepository));
     }
 
     private boolean hasForeignKeys(Person person) {
